rename, cleanup links
